+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 ----------------
 
 - Add tile option to specify filename of a multifile gist.
+- Render gist code in noscript tag.
 
 
 0.1 (2017-06-04)